Steiner Algorithm
The Steiner tree problem, motorway problem, or minimum Steiner tree
problem, named after Jakob Steiner, is a problem in combinatorial
optimization, which may be formulated in a number of settings, with the
common part being that it is required to find the shortest interconnect
for a given set of objects.

The Steiner tree problem is superficially similar to the minimum
spanning tree problem: given a set V of points (vertices), interconnect
them by a network (graph) of shortest length, where the length is the
sum of the lengths of all edges. The difference between the Steiner
tree problem and the minimum spanning tree problem is that, in the
Steiner tree problem, extra intermediate vertices and edges may be
added to the graph in order to reduce the length of the spanning tree.
These new vertices introduced to decrease the total length of
connection are known as Steiner points or Steiner vertices. It has been
proved that the resulting connection is a tree, known as the Steiner
tree. There may be several Steiner trees for a given set of initial
vertices.

The Steiner tree problem has applications in circuit layout or network
design. Most versions of the Steiner tree problem are NP-complete. In
fact, one of these was among Karp's original 21 NP-complete problems.
Some restricted cases can be solved in polynomial time. In practice,
heuristics are used.